src/embeddings-server/requirements.txt
chromadb
llama-index
python-dotenv
fastapi
uvicorn
pydantic
pydantic-settings
sentence-transformers
httpx